+++ android/src/main/kotlin/com/itisnajim/device_policy_controller/AppDeviceAdminReceiver.kt
@@ -1,29 +1,30 @@
 package com.itisnajim.device_policy_controller
 
-import android.app.ActivityManager
 import android.app.admin.DeviceAdminReceiver
 import android.app.admin.DevicePolicyManager
 import android.content.Context
 import android.content.Intent
 import android.preference.PreferenceManager
 import android.util.Log
+import android.view.KeyEvent
+import io.flutter.embedding.engine.FlutterEngine
+import io.flutter.embedding.engine.dart.DartExecutor
+
 
 class AppDeviceAdminReceiver : DeviceAdminReceiver() {
     companion object {
-        fun log(message: String) = Log.d("dpc::", message);
+        fun log(message: String) = Log.d("dpc::", message)
 
-        private const val KEY_SHOULD_START_ACTIVITY = "should_start_activity_at_boot"
+        private const val KEY_IS_FROM_BOOT_COMPLETED = "is_from_boot_completed"
 
-        fun shouldStartActivityAtBootCompleted(context: Context): Boolean {
+        fun setIsFromBootCompleted(context: Context, value: Boolean) {
             val sharedPreferences = PreferenceManager.getDefaultSharedPreferences(context)
-            return sharedPreferences.getBoolean(KEY_SHOULD_START_ACTIVITY, false)
+            sharedPreferences.edit().putBoolean(KEY_IS_FROM_BOOT_COMPLETED, value).apply()
         }
 
-        fun setShouldStartActivityAtBootCompleted(context: Context, shouldStart: Boolean) {
+        fun isFromBootCompleted(context: Context): Boolean {
             val sharedPreferences = PreferenceManager.getDefaultSharedPreferences(context)
-            val editor = sharedPreferences.edit()
-            editor.putBoolean(KEY_SHOULD_START_ACTIVITY, shouldStart)
-            editor.apply()
+            return sharedPreferences.getBoolean(KEY_IS_FROM_BOOT_COMPLETED, false)
         }
     }
 
@@ -32,7 +33,7 @@ class AppDeviceAdminReceiver : DeviceAdminReceiver() {
         val i: Intent? =
             context.packageManager.getLaunchIntentForPackage(context.packageName)
         if (i != null) {
-            i.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
+            i.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK)
             context.startActivity(i)
         } else {
             log("Couldn't start activity")
@@ -42,35 +43,21 @@ class AppDeviceAdminReceiver : DeviceAdminReceiver() {
     override fun onReceive(context: Context, intent: Intent) {
         super.onReceive(context, intent)
         val action = intent.action
+        val extras = intent.extras
+        log("onReceive: action: $action, extras: $extras")
         if (DevicePolicyManager.ACTION_MANAGED_PROFILE_PROVISIONED == action || Intent.ACTION_MANAGED_PROFILE_ADDED == action) {
             PreferenceManager.getDefaultSharedPreferences(context)
-                .edit().putBoolean("is_provisioned", true).apply();
+                .edit().putBoolean("is_provisioned", true).apply()
         }
-
-        val i: Intent? =
-            context.packageManager.getLaunchIntentForPackage(context.packageName)
-        val isActivityVisible = isActivityVisible(context);
-        val shouldStartActivityAtBootCompleted = shouldStartActivityAtBootCompleted(context)
-        log("shouldStartActivityAtBootCompleted: $shouldStartActivityAtBootCompleted")
-        if (action == Intent.ACTION_BOOT_COMPLETED && i != null && shouldStartActivityAtBootCompleted && !isActivityVisible) {
-            i.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-            context.startActivity(i)
-        }else if(action == Intent.ACTION_BOOT_COMPLETED && !isActivityVisible && shouldStartActivityAtBootCompleted){
-            log("Couldn't start activity")
+        if (action == Intent.ACTION_BOOT_COMPLETED) {
+            setIsFromBootCompleted(context, true)
+            val flutterEngine = FlutterEngine(context.applicationContext)
+            flutterEngine.plugins.get(DevicePolicyControllerPlugin::class.java) as DevicePolicyControllerPlugin?
+            flutterEngine.dartExecutor.executeDartEntrypoint(DartExecutor.DartEntrypoint.createDefault())
+            val channel =
+                DevicePolicyControllerPlugin.methodChannel(flutterEngine.dartExecutor.binaryMessenger)
+            channel.invokeMethod("handleBootCompleted", null)
         }
-
     }
-
-    private fun isActivityVisible(context: Context): Boolean {
-        val activityManager = context.getSystemService(Context.ACTIVITY_SERVICE) as ActivityManager
-        val tasks =
-            activityManager.runningAppProcesses.filter {
-                it.processName == context.packageName &&
-                        it.importance == ActivityManager.RunningAppProcessInfo.IMPORTANCE_FOREGROUND
-            }
-
-        return tasks.isNotEmpty();
-    }
-
 
 }