+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## 0.0.3
+
+New methods:
+* setAsLauncher
+* startApp
+* handleBootCompleted
+* isScreenAwake
+
+Methods to manage storing simple data in the device preferences:
+```dart
+Future<String?> get(String contentKey, {String? defaultContent});
+Future<void> put(String contentKey, {String? content});
+Future<void> remove(String contentKey);
+Future<void> clear();
+```
+
 ## 0.0.2
 
 Add methods: